fix: finalize runtime.lifecycle.stopped from runtime result on shutdown
Wait for run completion before canceling metrics sidecars so terminal lifecycle outcomes are not mislabeled as canceled during teardown.
